delete tweet + refactor code
